Flatten domain module
The domain module (dtfs) must be root and not a stereotype like "model" or "core" or "domain". Domain modules have the same name as the whole project, with no stereotypes, and in a case by case basis.

As a utopia, everything should be a domain (hence homogeneous) but that would be overengineering. Let at least the domain of your project/org/business be a proper domain.